+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 
 Unreleased is the current development version.
+- Add support for area selection with fldmean (Fldmean box selection #409)
 
 - Environment simplified, dependencies are now mostly on the pyproject file (A simpler environment.yml #286)
 - Intake esm functionality added back (Fix intake-esm #287)